requirements.txt
asgiref==3.7.2
async-timeout==4.0.2
azure-common==1.1.28
bcrypt==4.0.1
cachetools==5.3.1
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
contourpy==1.0.7
cycler==0.11.0
dash==2.8.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
distlib==0.3.6
Django==4.2.6
django-cors-headers==4.3.0
django-restframework==0.0.1
djangorestframework==3.14.0
dnspython==2.3.0
filelock==3.12.1
Flask==2.2.2
Flask-Bcrypt==1.0.1
Flask-Cors==3.0.10
Flask-MySQLdb==1.0.1
fonttools==4.39.0
future==0.18.3
google-api-core==2.12.0
google-api-python-client==2.104.0
google-auth==2.23.3
google-auth-httplib2==0.1.1
googleapis-common-protos==1.61.0
graphviz==0.20.1
greenlet==2.0.2
httplib2==0.22.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
kiwisolver==1.4.4
MarkupSafe==2.1.2
matplotlib==3.7.1
mysql==0.0.3
mysql-connector-python==8.0.32
mysqlclient==2.1.1
numpy==1.24.2
oauth2client==4.1.3
oauthlib==3.2.2
packaging==22.0
pandas==1.5.3
pdf2image==1.16.2
Pillow==9.4.0
platformdirs==3.5.3
plotly==5.13.1
protobuf==3.20.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pymongo==4.3.3
PyMySQL==1.0.2
pyparsing==3.0.9
PySocks==1.7.1
pytesseract==0.3.10
python-dateutil==2.8.2
python-dotenv==0.21.1
pytz==2022.7.1
redis==4.5.1
requests==2.28.2
rsa==4.9
schedule==1.2.0
scikit-learn==1.2.2
scipy==1.10.1
seaborn==0.12.2
six==1.16.0
sklearn==0.0.post1
spotipy==2.22.1
SQLAlchemy==2.0.3
sqlparse==0.4.4
tenacity==8.2.2
threadpoolctl==3.1.0
typing_extensions==4.4.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.13
virtualenv==20.23.0
Werkzeug==2.2.3
zope.event==4.6
zope.interface==5.5.2
